python-backend/requirements.txt
fastapi
uvicorn
openai-whisper
coqui-tts
torch
pydub
python-multipart
